requirements.txt
google-genai>=1.0.0
PyGithub>=2.1.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0

